Fixed UI elements not rendering at the correct places when resizing quickly.
This issue seems to be caused by the window only updating between two draw loop iterations while the UI calculations already use the updated width/height values. Thus, the actual window is one frame behind.
It was fixed by simply first updating the UI with the window values used in this iteration before updating the window for the next one.